Avoid scrolling when dragging over any block component
