+++ super_editor/test/super_editor/supereditor_scrolling_test.dart
@@ -280,7 +280,7 @@ void main() {
               ],
             ),
           )
-          .withAddedComponents([const _FakeImageComponentBuilder(size: editorSize)])
+          .withAddedComponents([const FakeImageComponentBuilder(size: editorSize)])
           .withEditorSize(editorSize)
           .pump();
 
@@ -306,38 +306,4 @@ void main() {
       );
     });
   });
-}
-
-/// A [ComponentBuilder] which builds an [ImageComponent] that always renders
-/// images as a [SizedBox] with the given [size].
-class _FakeImageComponentBuilder implements ComponentBuilder {
-  const _FakeImageComponentBuilder({
-    required this.size,
-  });
-
-  final Size size;
-
-  @override
-  SingleColumnLayoutComponentViewModel? createViewModel(Document document, DocumentNode node) {
-    return null;
-  }
-
-  @override
-  Widget? createComponent(
-      SingleColumnDocumentComponentContext componentContext, SingleColumnLayoutComponentViewModel componentViewModel) {
-    if (componentViewModel is! ImageComponentViewModel) {
-      return null;
-    }
-
-    return ImageComponent(
-      componentKey: componentContext.componentKey,
-      imageUrl: componentViewModel.imageUrl,
-      selection: componentViewModel.selection,
-      selectionColor: componentViewModel.selectionColor,
-      imageBuilder: (context, imageUrl) => SizedBox(
-        height: size.height,
-        width: size.width,
-      ),
-    );
-  }
 }
